requirements.txt
django
django-colorful
django-ratelimit
